Create action to fetch a single post for detail post page
